requirements.txt
Flask
openai
numpy
scikit-learn
paho-mqtt